move rerun functionality to rerun.ts
